+++ src/lib/tls/tls12/tls_handshake_state.cpp
@@ -254,6 +254,11 @@ Session_Ticket Handshake_State::session_ticket() const {
 std::unique_ptr<KDF> Handshake_State::protocol_specific_prf() const {
    const std::string prf_algo = ciphersuite().prf_algo();
 
+   auto user_defined_prf = m_callbacks.tls_protocol_specific_kdf("TLS-12-PRF", prf_algo);
+   if(user_defined_prf) {
+      return user_defined_prf;
+   }
+
    if(prf_algo == "MD5" || prf_algo == "SHA-1") {
       return KDF::create_or_throw("TLS-12-PRF(SHA-256)");
    }