+++ news.rst
@@ -19,6 +19,8 @@ Version 3.10.0, Not Yet Released
 
 * Fix various typos in the source and documentation (GH #5071)
 
+* Add optional callback to provide a user-defined TLS key derivation function
+
 Version 3.9.0, 2025-08-05
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^